add javadoc comments. fix some possible runtime errors

+++ src/main/java/com/library/services/QueryParser.java
@@ -7,11 +7,38 @@ import com.library.entities.Book;
 
 import java.sql.Connection;
 
+/**
+ * Parse queries
+ */
 public class QueryParser {
 
+    private static QueryParser queryParser = null;
     private BookDAO bookDAO = new BookDAOImpl();
-    private ConsoleDialog consoleDialog = new ConsoleDialog();
+    private ConsoleDialog consoleDialog = ConsoleDialog.getInstance();
 
+    /**
+     * @return instance of QueryParser
+     */
+    public static QueryParser getInstance() {
+        if (queryParser == null) {
+            return new QueryParser();
+        } else {
+            return queryParser;
+        }
+    }
+
+    private QueryParser() {
+
+    }
+
+    /**
+     * Parse queries. Call specified methods
+     * Invoke {@link #parseCommand(String)}, {@link #parseAuthor(String)}, {@link #parseName(String)}
+     *
+     * @param query      user's query
+     * @param connection database connection
+     * @throws Exception
+     */
     public void parseQuery(String query, Connection connection) throws Exception {
 
         String command = parseCommand(query);
@@ -31,39 +58,64 @@ public class QueryParser {
                     consoleDialog.writeFail();
                 }
             }
+            return;
         }
         if (command.toLowerCase().equals(Commands.UPDATE.toString())) {
-            Book book = consoleDialog.getElaboration(bookDAO.getBooksByName(connection, name));
-            book.setName(consoleDialog.getNewName());
-            if (bookDAO.updateBook(connection, book) > 0) {
-                consoleDialog.writeUpdate(book);
+            Book book = consoleDialog.getSpecificaton(bookDAO.getBooksByName(connection, name));
+            if (book != null) {
+                book.setName(consoleDialog.getNewName());
+                if (bookDAO.updateBook(connection, book) > 0) {
+                    consoleDialog.writeUpdate(book);
+                }
             } else {
                 consoleDialog.writeFail();
             }
+            return;
         }
         if (command.toLowerCase().equals(Commands.REMOVE.toString())) {
-            Book book = consoleDialog.getElaboration(bookDAO.getBooksByName(connection, name));
-            if (bookDAO.deleteBook(connection, book) > 0) {
-                consoleDialog.writeDeletion(book);
+            Book book = consoleDialog.getSpecificaton(bookDAO.getBooksByName(connection, name));
+            if (book != null) {
+                if (bookDAO.deleteBook(connection, book) > 0) {
+                    consoleDialog.writeDeletion(book);
+                }
             } else {
                 consoleDialog.writeFail();
             }
+            return;
         }
         if (!query.equals("")) {
             consoleDialog.writeNoCommand();
         }
     }
 
+    /**
+     * Parse command value from user's query
+     *
+     * @param query user's query
+     * @return parsed command value
+     */
     private String parseCommand(String query) {
         return query.split(" ")[0];
     }
 
+    /**
+     * Parse name value from user's query
+     *
+     * @param query user's query
+     * @return parsed name value
+     */
     private String parseName(String query) {
         int start = query.indexOf("\"");
         int end = query.lastIndexOf("\"");
         return start != -1 && end != -1 ? query.substring(start + 1, end).trim() : "";
     }
 
+    /**
+     * Parse author value from user's query
+     *
+     * @param query user's query
+     * @return parsed author value
+     */
     private String parseAuthor(String query) {
         int end = query.indexOf("\"");
         return end != -1 ? query.substring(Commands.ADD.toString().length(), end).trim() : "";
